+++ astropy/io/fits/tests/test_fitsinfo.py
@@ -5,7 +5,7 @@ import pytest
 from astropy import __version__ as version
 from astropy.io.fits.scripts import fitsinfo
 
-from . import FitsTestCase
+from .conftest import FitsTestCase
 
 
 class TestFitsinfo(FitsTestCase):
