+++ astropy/io/fits/tests/test_groups.py
@@ -9,7 +9,7 @@ import pytest
 from astropy.io import fits
 from astropy.utils.misc import _NOT_OVERWRITING_MSG_MATCH
 
-from . import FitsTestCase, home_is_temp
+from .conftest import FitsTestCase
 from .test_table import comparerecords
