+++ astropy/io/fits/tests/test_header.py
@@ -16,7 +16,7 @@ from astropy.io.fits.verify import VerifyError, VerifyWarning
 from astropy.utils.exceptions import AstropyUserWarning
 from astropy.utils.misc import _NOT_OVERWRITING_MSG_MATCH
 
-from . import FitsTestCase, home_is_temp
+from .conftest import FitsTestCase
 
 
 def test_shallow_copy():
