+++ astropy/io/fits/tests/test_uint.py
@@ -8,7 +8,7 @@ from numpy.testing import assert_array_equal
 
 from astropy.io import fits
 
-from . import FitsTestCase
+from .conftest import FitsTestCase
 
 
 class TestUintFunctions(FitsTestCase):
